+++ bin/actions/inject_dependency_reference_action.js
@@ -2,6 +2,7 @@ const ApplicationConfiguration = require('../configurations/application')
 const ManifestConfiguration = require('../configurations/manifest')
 const ReferenceParser = require(`../parsers/reference_parser`)
 const StatusLog = require('../support/status_log')
+const _ = require('lodash')
 
 const injectDependencyReferenceAction = (reference, options) => {
   return new Promise((resolve, reject) => {
@@ -9,7 +10,7 @@ const injectDependencyReferenceAction = (reference, options) => {
     const activeDependency = (options['save-dev']) ? manifestConfiguration.devDependencies() : manifestConfiguration.dependencies()
 
     if (reference) {
-      const manifestVersionSeparator = manifestConfiguration.system.configurationSystem.versionSeparator
+      const manifestVersionSeparator = manifestConfiguration.system.versionSeparator
       const versionSeparators = ApplicationConfiguration.get(`rules.patternMarkers.version`)
       const { installedName } = ReferenceParser.referenceToArchiveRequest(reference, options.rename)