+++ bin/actions/upload_archive_action.js
@@ -1,14 +1,23 @@
 const StatusLog = require('../support/status_log')
 
 const uploadArchiveAction = (resourceBundle, manifestConfiguration) => {
-  const { archiveBundle, TransitTool } = resourceBundle
+  return new Promise((resolve, reject) => {
+    const { archiveBundle, TransitTool } = resourceBundle
 
-  return TransitTool
-    .push(archiveBundle, manifestConfiguration)
-    .then(() => {
-      StatusLog.notify(`uploaded`, archiveBundle.uuid)
-      return resourceBundle
-    })
+    if (!archiveBundle.releaseAsset) {
+      reject(new Error(`you must specify the asset to publish`))
+    }
+
+    TransitTool
+      .push(archiveBundle, manifestConfiguration)
+      .then(() => {
+        StatusLog.notify(`uploaded`, archiveBundle.uuid)
+        resolve(resourceBundle)
+      })
+      .catch(err => {
+        reject(err)
+      })
+  })
 }
 
 module.exports = uploadArchiveAction