+++ bin/configurations/application.js
@@ -6,7 +6,6 @@ const os = require('os')
 
 class ApplicationConfiguration {
   constructor () {
-
     this.__configuration = {}
     this.__paths = {
       application: `configuration`,
@@ -21,7 +20,6 @@ class ApplicationConfiguration {
 
   initialize () {
     const currentFolder = process.cwd()
-    const homeFolder = os.homedir()
 
     let baseFolder = ''
     const parentPaths = _.map(_.compact(currentFolder.split(path.sep)), (aspect) => {
@@ -29,16 +27,27 @@ class ApplicationConfiguration {
       return baseFolder
     })
 
-    const paths = _([
+    _([
       this.__paths.application,
+      this.__paths.global,
       ...parentPaths,
-      this.__paths.global
+      this.__paths.local
     ])
     .uniq()
-    .map((configurationPath) => `${configurationPath}${path.sep}.vaultrc`)
-    .each((configurationPath) => this.loadFile(configurationPath))
+    .map((configurationPath) => this.__vaultRCFile(configurationPath))
+    .each((configurationPath) => this.__loadFile(configurationPath))
 
     return this
+  }
+
+  initializeLocal () {
+    if (!this.hasLocalFile()) {
+      this.saveLocalFile({
+        sources: {},
+        paths: {},
+        rules: {}
+      })
+    }
   }
 
   saveGlobalFile (settings) {
@@ -50,7 +59,7 @@ class ApplicationConfiguration {
   }
 
   saveFile (type, settings) {
-    const filename = this.__paths[type]
+    const filename = this.__vaultRCFile(this.__paths[type])
     const configurationContent = this.__getFile(filename)
 
     return this._saveFile(filename, _.merge({}, configurationContent, settings))
@@ -61,9 +70,21 @@ class ApplicationConfiguration {
     return this
   }
 
-  loadFile (filename) {
+  hasLocalFile () {
+    return fs.existsSync(this.__vaultRCFile(this.__paths['local']))
+  }
+
+  loadFile (type) {
+    return this.__loadFile(this.__vaultRCFile(this.__paths[type]))
+  }
+
+  __loadFile (filename) {
     this.__configuration = _.merge(this.__configuration, this.__getFile(filename))
     return this
+  }
+
+  __vaultRCFile (configurationPath) {
+    return `${configurationPath}${path.sep}.vaultrc`
   }
 
   __getFile (filename) {