+++ bin/servants/security_servant.js
@@ -7,12 +7,17 @@ class SecurityServant {
    * @param {String} message
    */
   static encrypt (message) {
-    const cipher = crypto.createCipher(
-      `aes-256-ctr`,
-      this.__salt()
-    )
+    try {
+      const cipher = crypto.createCipher(
+        `aes192`,
+        this.__salt()
+      )
 
-    return cipher.update(message, 'utf8', 'hex') + cipher.final('hex')
+      return cipher.update((message + '').toString(), 'utf8', 'hex') + cipher.final('hex')
+    } catch (e) {
+      console.error(`encrypt`, e)
+      return ''
+    }
   }
 
   /**
@@ -20,12 +25,17 @@ class SecurityServant {
    * @param {String} cipher
    */
   static decrypt (cipher) {
-    const decipher = crypto.createDecipher(
-      `aes-256-ctr`,
-      this.__salt()
-    )
+    try {
+      const decipher = crypto.createDecipher(
+        `aes192`,
+        this.__salt()
+      )
 
-    return decipher.update(cipher, 'hex', 'utf8') + decipher.final('utf8')
+      return decipher.update(cipher, 'hex', 'utf8') + decipher.final('utf8')
+    } catch (e) {
+      console.error(`decrypt`, e)
+      return ''
+    }
   }
 
   static __salt () {
@@ -33,7 +43,7 @@ class SecurityServant {
       gid
     } = os.userInfo()
 
-    return `${gid}`
+    return `__${gid}`
   }
 }