+++ changelog.md
@@ -1,5 +1,19 @@
 # Klepto Changelog
 
+## 0.9.0
+
+- ADDED `--encrypt` to `configure` to encrypt values on .vaultrc
+- ADDED `--all` to `uninstall` to erase the entire vault/ folder
+- REMOVED `status` command, it was not implemented and mostly is implemented through updates to `configure`
+- ADDED documentation
+- ADDED `configuration` will show the entire configuration object affecting a project folder
+- ADDED `initialize` creates a configuration file (`.vaultrc`) along with a resource file (`vault.json`)
+- FIXED always include the local folder last and the global folder before the folder traversal when calculating the application configuration
+- FIXED make sure the global configuration is before any folder traversal configurations
+- FIXED save klepto configuration
+- FIXED `uninstall` will fail if the archive does not exist
+- FIXED publish FTP will report errors
+
 ## 0.8.0
 
 - ADDED `klepto publish` to prepare a build an archive and upload to a host